backend/requirements.txt
# Enterprise AI Security Red Teaming Platform
# Backend Requirements

# FastAPI and Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Async Job Queue
redis>=5.0.0
rq>=1.15.0

# HTTP Client for Model APIs
httpx>=0.26.0
aiohttp>=3.9.0

# Model Provider SDKs
openai>=1.10.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
slowapi>=0.1.9

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Logging and Monitoring
structlog>=23.2.0

# Data Validation
email-validator>=2.1.0

# PDF Generation
reportlab>=4.0.0

# CORS
starlette>=0.35.0
